requirements.txt
aiohttp==3.7.3
aioice==0.7.3
aiortc==1.1.1
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autobahn==21.2.1
av==8.0.3
beautifulsoup4==4.6.0
black==20.8b1
blinker==1.4
Brlapi==0.6.6
callee
certifi==2018.1.18
cffi==1.14.4
chardet==3.0.4
click==7.1.2
crc32c==2.2
cryptography==3.4.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dnspython==2.1.0
feedparser==5.2.1
flake8==3.9.0
Flask==1.1.2
future==0.18.2
futures==3.1.1
gast==0.2.2
gevent==21.1.2
geventhttpclient==1.4.4
graphsurgeon==0.4.5
graphviz==0.16
greenlet==1.0.0
grpcio==1.35.0
grpcio-tools==1.35.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==21.0.0
idna==2.6
idna-ssl==1.1.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jetson.GPIO==2.0.11
Jinja2==2.11.3
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.3.4
mccabe==0.6.1
mock==3.0.5
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
netifaces==0.10.9
networkx==2.5
numpy==1.16.1
oauth==1.0.1
oauthlib==2.0.6
onboard==1.4.1
packaging==20.9
PAM==0.4.2
pandas==1.1.5
pathspec==0.8.1
pbr==5.5.1
Pillow==8.1.0
playsound==1.2.2
pluggy==0.13.1
protobuf==3.14.0
psutil==5.8.0
py==1.10.0
pybind11==2.6.2
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyee==8.1.0
pyflakes==2.3.0
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pylibsrtp==0.6.7
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyserial==3.5
pytest==6.2.2
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-rapidjson==1.0
pytz==2018.3
PyWavelets==1.1.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
regex==2020.11.13
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorrt==7.1.3.0
testresources==2.0.1
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
torch @ file:///home/gruppe38/torch-1.7.0-cp36-cp36m-linux_aarch64.whl
torchvision==0.2.2.post3
tqdm==4.56.0
tritonclient @ file:///home/gruppe38/clients/python/tritonclient-2.6.0-py3-none-linux_aarch64.whl
txaio==21.2.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
websockets==8.1
Werkzeug==1.0.1
xkit==0.0.0
yarl==1.6.3
youtube-dl==2018.3.14
zipp==3.4.1
zope.event==4.5.0
zope.interface==4.3.2
